move histogram types to util file
