+++ src/app/gene-scores/gene-scores.state.ts
@@ -1,10 +1,7 @@
 import { createReducer, createAction, on, props, createFeatureSelector } from '@ngrx/store';
 import { cloneDeep } from 'lodash';
 import { reset } from 'app/users/state-actions';
-import { CategoricalHistogramView } from 'app/genomic-scores-block/genomic-scores-block';
-
-
-export type HistogramType = 'continuous' | 'categorical';
+import { CategoricalHistogramView, HistogramType } from 'app/utils/histogram-types';
 
 export interface GeneScoresState {
   histogramType: HistogramType;
